Added war-based archetype, fixed grizzly archetype.
Added archetype for servlet-based applications using war packaging.
For now the archetype uses deployment descriptor. Once the ResourceConfig/JerseyApplication
are refactored, need to change it to use the application class instead.

For Grizzly artifact had to change Application to JerseyApplication as seems it was missed
during the refactoring. Also made a minor change to the name of the grizzly archetype.

Made both artifacts consistent and did some more cleanups.
Ouch, and mised deleting MainTest in the previous patchset. Fixed now.

Change-Id: I5c7314e0381b11493b3f6ef721cd8a6606c93ff8