Changed the key for the MessageBodyReader cache.
The reason for this change is that there is a cahe of MessageBodyReaders that grows until memory runs out.
This happens when getting lots of multipart messages (with unique boundaries). There is a cache that
eventually will fill up the available memory because boundaries make the keys unique and new entries
are added to the cache for each request.

This change will create a new key for the cache that doesn't contain the additional parameters and won't
make for example the multipart/mixed media type uniqe just because the boundaries are uniqe.