JERSEY-2347 Update release process to use 1.7 target by default - part I.
The global source level was changed to 1.7 and a new profile 'jdk6' was created in core-common and core-client.
Calling mvn clean package -Pjdk6 in client/common creates a jar with the -jdk6 classifier built with 1.6 source level.
This is the first part of the task, the second part will take place on the Hudson side.

Change-Id: I22ff6969f98bb638e06b96acd6f26b708db5e279