JERSEY-2651 ProviderBinder.bindProvider must be called, even if it is bound by a component provider
Components bound by component providers prevent them from being registered and available as JAX-RS components. The concrete example here is `ContainerRequestFilter` classes, as seen in the test in the spring3 example - when it is annotated with `@Component`, it is bound by `SpringComponentProvider.bind`, and this prevents the filter from actually running on incoming requests. Passing the class on to `ProviderBinder.bindProvider` fixes this.