A minor change to SSE Item Store Example.
This change adds a new broadcaster listener in attempt to uncover a problem with the intermittently failing ItemStoreResourceTest.testEventSourceReconnect test.

The observed behavior is that sometimes an empty unnamed event arrives to the client, when a “round-N-NN” event is expected instead. We suspect that there may be a potential race condition in the example between broadcaster.closeAll() and broadcaster.broadcast(…) invocations induced by the test, which is causing that an event output gets closed before the broadcasting of the last event is finished. Any such race condition should be now caught as an exception event and logged by the new listener.

Change-Id: If9f04a1bd0bed6c086ffb5fd9004a14c0f2e6498
Signed-off-by: Marek Potociar <marek.potociar@oracle.com>