JERSEY-3023: Client proxy not marshalling object to json due to order of arguments
The original algorithm  loops through all the method parameters
and checks if the parameter contains an @XxxParam annotation. If it doesn't,
then it is determined to be the entity parameter. The problem with this is that
it doesn't take into account @Context and other possible annotation
that are not the entity, so the entity needs to always be the last parameter
in the event there are other parameters without @XxxParam.

There's not really much you can do about this, as the possibilities are limitless
as to what could cause a fault in this algorithm. The only reasonable solution
(or maybe work around) is to just allow the developer to specify the entity
index explicitly in an overloaded factory method. Without using the overload,
this fix leaves the program to proceed as it normally would, so there is no 
effect to any previous usages.

Tests were added to verify the behavior of the entity index, in the test class
WebResourceFactoryEntityIndexTest