Add callback hook for modifying the invocation.
This is intented to be used when dealing with APIs which use the entity
body to pass metadata.  As an example, json-rpc (http://json-rpc.org/)
specifies a protocol where the url remains constant, and a paramater
"method" determines the action performed.  That means you can't leverage
types because you end up an API that looks like this...

public class FreeIpaClient {
  @Consumes("application/json")
  @Produces("application/json")
  @POST @Path("/ipa/sessions/json")
  Response doRpc(JsonRpcCall call);
}

All leverage is lost.  In order to build a tool which is able to execute
different json-rpc actions based on the interface, there needs to be
some scope where the invocation context and the invoked method are
available.  This change creates that scope.  You can now have the
following.

public class FreeIpaClient {
  @Consumes("application/json")
  @Produces("application/json")
  @POST @Path("/ipa/sessions/json")
  @CustomJsonRpcAnno("user_show")
  UserShowResponse userShow(UserShowParams params);

  @Consumes("application/json")
  @Produces("application/json")
  @POST @Path("/ipa/sessions/json")
  @CustomJsonRpcAnno("user_delete")
  UserDeleteResponse userDelete(UserDeleteParams params);
}

With this interface, you can set a property on the Invocation.Builder
which contains the annotation information from the invoked method
userShow or userDelete.

You still need to do work to register a message body write/reader to
append the metadata contained in the annotation.  If desired you could
even do reflection based off of the method name.